Import ACPICA 20150408.

+++ source/tools/acpiexec/aeregion.c
@@ -687,8 +687,8 @@ AeRegionHandler (
                 {
                     /* Right overlap */
 
-                    RegionElement->Length = BaseAddress -
-                        RegionAddress + Length;
+                    RegionElement->Length = (UINT32) (BaseAddress -
+                        RegionAddress + Length);
                     BufferResize = TRUE;
                 }
 
@@ -699,8 +699,8 @@ AeRegionHandler (
                     /* Left overlap */
 
                     RegionElement->Address = BaseAddress;
-                    RegionElement->Length = RegionAddress -
-                        BaseAddress + RegionElement->Length;
+                    RegionElement->Length = (UINT32) (RegionAddress -
+                        BaseAddress + RegionElement->Length);
                     BufferResize = TRUE;
                 }
